backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
redis==5.0.1
scikit-learn==1.6.1
pandas==2.1.4
joblib==1.3.2
python-dotenv==1.0.0
requests==2.31.0
werkzeug==3.0.1
numpy==1.26.2
gunicorn==21.2.0
bcrypt==4.1.2
PyJWT==2.8.0